+++ dto/pushed_auth_req_action.go
@@ -1,0 +1,28 @@
+//
+// Copyright (C) 2020 Authlete, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+// either express or implied. See the License for the specific
+// language governing permissions and limitations under the
+// License.
+
+package dto
+
+type PushedAuthReqAction string
+
+const (
+	PushedAuthReqAction_CREATED               = PushedAuthReqAction(`CREATED`)
+	PushedAuthReqAction_BAD_REQUEST           = PushedAuthReqAction(`BAD_REQUEST`)
+	PushedAuthReqAction_UNAUTHORIZED          = PushedAuthReqAction(`UNAUTHORIZED`)
+	PushedAuthReqAction_FORBIDDEN             = PushedAuthReqAction(`FORBIDDEN`)
+	PushedAuthReqAction_PAYLOAD_TOO_LARGE     = PushedAuthReqAction(`PAYLOAD_TOO_LARGE`)
+	PushedAuthReqAction_INTERNAL_SERVER_ERROR = PushedAuthReqAction(`INTERNAL_SERVER_ERROR`)
+)